# contracts/web-api.openapi.yaml
openapi: 3.1.0
info:
  title: RayleaBot Web API
  version: 0.1.0
  description: >
    Fixture-ready HTTP contract. This file intentionally contains only
    the currently frozen management routes.
...
paths:
...
  /api/config:
    get:
      summary: Query the current effective management configuration snapshot.
      operationId: getConfig
      security:
        - BearerAuth: []
        - CookieAuth: []
      responses:
        '200':
          description: Current effective config snapshot with sensitive fields redacted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigSnapshotResponse'
        '401':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
    put:
      summary: Validate and persist a new management configuration snapshot.
      operationId: updateConfig
      security:
        - BearerAuth: []
        - CookieAuth: []
          CSRFAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigDocument'
      responses:
        '200':
          description: Config written successfully. restart_required indicates whether a restart is still needed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigUpdateResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        Session token issued by /api/session/login or /api/setup/admin,
        passed as a Bearer token in the Authorization header.
    CookieAuth:
      type: apiKey
      in: cookie
      name: raylea_session
      description: Host-only HttpOnly SameSite=Strict browser session cookie. Secure is required when web.public_origin is HTTPS.
    CSRFAuth:
      type: apiKey
      in: header
      name: X-Raylea-CSRF
      description: Required on every state-changing request authenticated by CookieAuth. Invalid or missing values map to permission.denied.
    SetupTokenAuth:
      type: apiKey
      in: header
      name: X-Raylea-Setup-Token
      description: One-time 256-bit setup token. Invalid, missing, or reused values map to permission.denied.
    LauncherControlAuth:
      type: apiKey
      in: header
      name: X-Raylea-Launcher-Control
      description: Per-process launcher control token. Invalid or missing values map to permission.denied.
...
  responses:
    Error:
      description: Standard structured error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, message_key, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            message_key:
              type: string
            request_id:
              type: string
            details:
              type: object
              additionalProperties: true
...
    ConfigDocument:
      $ref: './config.user.schema.json'
    ConfigSnapshotResponse:
      type: object
      additionalProperties: false
      required: [config, effective_timezone]
      properties:
        effective_timezone:
          type: string
          minLength: 1
          description: IANA timezone currently used by the running scheduler and management time display. Remains unchanged after saving a timezone change until the server restarts. Defaults to Asia/Shanghai.
        config:
          $ref: '#/components/schemas/ConfigDocument'
        redacted_fields:
          type: array
          items:
            type: string
    ConfigApplyEffects:
      type: object
      additionalProperties: false
      required: [applied_now, reloaded_now, restart_required_fields]
      properties:
        applied_now:
          type: array
          items:
            type: string
        reloaded_now:
          type: array
          items:
            type: string
        restart_required_fields:
          type: array
          items:
            type: string
    ConfigUpdateResponse:
      type: object
      additionalProperties: false
      required: [config, effective_timezone, restart_required, apply_effects]
      properties:
        effective_timezone:
          type: string
          minLength: 1
          description: IANA timezone currently used by the running scheduler and management time display, including while config.scheduler.timezone awaits restart.
        config:
          $ref: '#/components/schemas/ConfigDocument'
        redacted_fields:
          type: array
          items:
            type: string
        restart_required:
          type: boolean
          description: Whether a server restart is required before the new config takes effect.
        apply_effects:
          $ref: '#/components/schemas/ConfigApplyEffects'
...